# server/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/shop/CreateShopOrder": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Shop"
                ],
                "summary": "用户申请创建订单",
                "parameters": [
                    {
                        "description": "订单信息",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.CreateOrder"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "订单创建成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/response.CreateShopOrderResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/shop/DeleteShopOrder": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Shop"
                ],
                "summary": "用户申请取消订单",
                "parameters": [
                    {
                        "description": "订单取消请求体",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.DeleteOrder"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "订单删除成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "msg": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/shop/GetUserOrders": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Shop"
                ],
                "summary": "用户查询其拥有的所有订单",
                "parameters": [
                    {
                        "description": "查询用户订单请求体",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.GetUserOrders"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "订单查询成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/response.GetUserOrdersResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "request.CreateOrder": {
            "type": "object",
            "required": [
                "price",
                "serviceType"
            ],
            "properties": {
                "createTime": {
                    "description": "订单创建时间",
                    "type": "integer",
                    "example": 1645084800
                },
                "expireTime": {
                    "description": "订单过期时间",
                    "type": "integer",
                    "example": 1676620800
                },
                "price": {
                    "description": "订单价格",
                    "type": "number",
                    "example": 99.99
                },
                "serviceType": {
                    "description": "服务类型：1-包年 2-包月",
                    "type": "integer",
                    "example": 1
                },
                "status": {
                    "description": "订单状态：0-待支付 1-已支付 2-已取消",
                    "type": "integer",
                    "example": 0
                }
            }
        },
...
        "request.DeleteOrder": {
            "type": "object",
            "required": [
                "orderId"
            ],
            "properties": {
                "orderId": {
                    "description": "订单ID",
                    "type": "integer"
                }
            }
        },
...
        "request.GetUserOrders": {
            "type": "object",
            "required": [
                "userId"
            ],
            "properties": {
                "userId": {
                    "description": "用户ID",
                    "type": "integer"
                }
            }
        },
...
        "response.CreateShopOrderResponse": {
            "type": "object",
            "properties": {
                "shoporder": {
                    "$ref": "#/definitions/shop.ShopOrder"
                }
            }
        },
        "response.GetUserOrdersResponse": {
            "type": "object",
            "properties": {
                "shoporders": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/shop.ShopOrder"
                    }
                },
                "total": {
                    "type": "integer"
                }
            }
        },
...
        "response.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "msg": {
                    "type": "string"
                }
            }
        },
...
        "shop.ShopOrder": {
            "type": "object",
            "properties": {
                "ID": {
                    "description": "主键ID",
                    "type": "integer"
                },
                "createTime": {
                    "type": "integer"
                },
                "createdAt": {
                    "description": "创建时间",
                    "type": "string"
                },
                "expireTime": {
                    "type": "integer"
                },
                "orderId": {
                    "type": "string"
                },
                "price": {
                    "type": "number"
                },
                "serviceType": {
                    "type": "integer"
                },
                "status": {
                    "type": "integer"
                },
                "updatedAt": {
                    "description": "更新时间",
                    "type": "string"
                },
                "userId": {
                    "type": "integer"
                }
            }
        },
...
    }
}